Switch to codespaces-linux from microsoft vscode-dev-containers

+++ .devcontainer/library-scripts/docker-debian.sh
@@ -4,7 +4,7 @@
 # Licensed under the MIT License. See https://go.microsoft.com/fwlink/?linkid=2090316 for license information.
 #-------------------------------------------------------------------------------------------------------------
 
-# Syntax: ./docker-debian.sh <enable non-root docker socket access flag> <source socket> <target socket> <non-root user>
+# Syntax: ./docker-debian.sh [enable non-root docker socket access flag] [source socket] [target socket] [non-root user]
 
 ENABLE_NONROOT_DOCKER=${1:-"true"}
 SOURCE_SOCKET=${2:-"/var/run/docker-host.sock"}
@@ -18,9 +18,6 @@ if [ "$(id -u)" -ne 0 ]; then
     exit 1
 fi
 
-# Ensure apt is in non-interactive to avoid prompts
-export DEBIAN_FRONTEND=noninteractive
-
 # Function to run apt-get if needed
 apt-get-update-if-needed()
 {
@@ -32,14 +29,19 @@ apt-get-update-if-needed()
     fi
 }
 
+# Ensure apt is in non-interactive to avoid prompts
+export DEBIAN_FRONTEND=noninteractive
+
+# Install apt-transport-https, curl, lsb-release, gpg if missing
+if ! dpkg -s apt-transport-https curl ca-certificates lsb-release > /dev/null 2>&1 || ! type gpg > /dev/null 2>&1; then
+    apt-get-update-if-needed
+    apt-get -y install --no-install-recommends apt-transport-https curl ca-certificates lsb-release gnupg2 
+fi
+
 # Install Docker CLI if not already installed
 if type docker > /dev/null 2>&1; then
     echo "Docker CLI already installed."
 else
-    if ! type curl > /dev/null 2>&1; then
-        apt-get-update-if-needed
-        apt-get -y install --no-install-recommends apt-transport-https ca-certificates curl gnupg2 lsb-release
-    fi
     curl -fsSL https://download.docker.com/linux/$(lsb_release -is | tr '[:upper:]' '[:lower:]')/gpg | (OUT=$(apt-key add - 2>&1) || echo $OUT)
     echo "deb [arch=amd64] https://download.docker.com/linux/$(lsb_release -is | tr '[:upper:]' '[:lower:]') $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list
     apt-get update
@@ -50,6 +52,7 @@ fi
 if type docker-compose > /dev/null 2>&1; then
     echo "Docker Compose already installed."
 else
+
     LATEST_COMPOSE_VERSION=$(curl -sSL "https://api.github.com/repos/docker/compose/releases/latest" | grep -o -P '(?<="tag_name": ").+(?=")')
     curl -sSL "https://github.com/docker/compose/releases/download/${LATEST_COMPOSE_VERSION}/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
     chmod +x /usr/local/bin/docker-compose
@@ -75,8 +78,10 @@ fi
 
 # If enabling non-root access and specified user is found, setup socat and add script
 chown -h "${USERNAME}":root "${TARGET_SOCKET}"        
-apt-get-update-if-needed
-apt-get -y install socat
+if ! dpkg -s socat > /dev/null 2>&1; then
+    apt-get-update-if-needed
+    apt-get -y install socat
+fi
 tee /usr/local/share/docker-init.sh > /dev/null \
 << EOF 
 #!/usr/bin/env bash
@@ -141,7 +146,8 @@ fi
 # Execute whatever commands were passed in (if any). This allows us 
 # to set this script to ENTRYPOINT while still executing the default CMD.
 set +e
-"\$@"
+exec "\$@"
 EOF
 chmod +x /usr/local/share/docker-init.sh
 chown ${USERNAME}:root /usr/local/share/docker-init.sh
+echo "Done!"
